+++ src/stop.sh
@@ -1,7 +1,11 @@
 #!bin/bash
 
-MASTER_PID=`ps xf | grep "python master/master.py" | awk '{print $1}'`
-WORKER_PID=`ps xf | grep "python worker/worker.py" | awk '{print $1}'`
-kill -9 $MASTER_PID
-kill -9 $WORKER_PID
+MASTER_PID=`ps xf | grep "python master/master.py" | grep -v grep | awk '{print $1}'`
+WORKER_PID=`ps xf | grep "python worker/worker.py" | grep -v grep | awk '{print $1}'`
+if [ -n "$MASTER_PID" ]; then
+    kill -9 $MASTER_PID
+fi
+if [ -n "$WOKER_PID" ]; then
+    kill -9 $WORKER_PID
+fi
